0.2.5
Improvements:
Specify einsum path to optimize the einsum computation, which increases the computation by 30%.
Added the option to input lmaxgrid in SH_deriv_store and Displacement_strains, which will control the output gridsizes for the SH derivatives and strain.

Corrections:
Minor corrections to B1986_nmax where the drho_corr to the tangential displacement was wrongly incrementally increased in the case where the density contrast is located in the surface topo or crust.
Fixed default values of base_drho and top_drho to be consistent in all functions.
Corrected a typo in SH_deriv and SH_deriv_store that caused a blow up of the derivatives at the pole.