0.2.6
The last sympy version now supports python > 3.8, so we have removed the requirement that sympy <= 1.7 should be used.

B1986_nmax.py:
Added a minor correction to the Banerdt system of equation in order for the degree-1 geoids (surface & moho) be always 0, as required in a center of mass reference frame. Previously, the degree-1 moho geoid had signals of a few meters.
Added an error message when the system is non-linear.