0.5.0
- Added a new function *Displacement_strains_shtools*.

This function is roughly similar to *Displacement_strains* expect that it uses Shtools to compute the spherical harmonic derivative.

This function is significantly faster than *Displacement_strains* and is recommended for most usage.

Note: There seem to be some slight ringing close to the poles with the current version of *Displacement_strains_shtools*. I'm not sure how to fix this. Nevertheless, I recommend always using this function, unless strains are computed at the pole.

- Examples updated to use *Displacement_strains_shtools*.
- Minor aesthetic changes to the figures.